Refactor move request definitions to it's own file
Temporal duplication of definitions for client and server
