+++ client/src/constants.ts
@@ -2,7 +2,3 @@ export namespace Constants {
   export const NATIVE_WORKFLOW_LANGUAGE_ID = "galaxyworkflow";
   export const CLEAN_WORKFLOW_DOCUMENT_SCHEME = "galaxy-clean-workflow";
 }
-
-export namespace LSRequestIdentifiers {
-  export const CLEAN_WORKFLOW = "galaxy-workflows-ls.cleanWorkflow";
-}
